+++ src/main/java/kr/tatine/manibogo_oms_v2/order/command/application/service/PlaceOrderService.java
@@ -2,7 +2,7 @@ package kr.tatine.manibogo_oms_v2.order.command.application.service;
 
 import kr.tatine.manibogo_oms_v2.ValidationErrorException;
 import kr.tatine.manibogo_oms_v2.common.ValidationError;
-import kr.tatine.manibogo_oms_v2.common.model.Address;
+import kr.tatine.manibogo_oms_v2.region.command.domain.Address;
 import kr.tatine.manibogo_oms_v2.common.model.Money;
 import kr.tatine.manibogo_oms_v2.common.model.Option;
 import kr.tatine.manibogo_oms_v2.common.model.PhoneNumber;
