--- src/main/java/kr/tatine/manibogo_oms_v2/order/query/dao/RegionDao.java
@@ -1,11 +1,0 @@
-package kr.tatine.manibogo_oms_v2.order.query.dao;
-
-import kr.tatine.manibogo_oms_v2.order.query.dto.RegionDto;
-
-import java.util.List;
-
-public interface RegionDao {
-
-    List<RegionDto> findDistinctAll();
-
-}
