+++ src/main/java/kr/tatine/manibogo_oms_v2/region/infra/JpqlRegionDao.java
@@ -1,8 +1,8 @@
-package kr.tatine.manibogo_oms_v2.order.infra;
+package kr.tatine.manibogo_oms_v2.region.infra;
 
 import jakarta.persistence.EntityManager;
-import kr.tatine.manibogo_oms_v2.order.query.dao.RegionDao;
-import kr.tatine.manibogo_oms_v2.order.query.dto.RegionDto;
+import kr.tatine.manibogo_oms_v2.region.query.dao.RegionDao;
+import kr.tatine.manibogo_oms_v2.region.query.dto.RegionDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
 
@@ -18,10 +18,8 @@ public class JpqlRegionDao implements RegionDao {
     public List<RegionDto> findDistinctAll() {
 
         final String query = """
-            SELECT DISTINCT new kr.tatine.manibogo_oms_v2.order.query.dto.RegionDto(r.sido, r.sigungu) FROM Region r
+            SELECT DISTINCT new kr.tatine.manibogo_oms_v2.region.query.dto.RegionDto(r.sido, r.sigungu) FROM ZipCodeRegion r
         """;
-
         return em.createQuery(query, RegionDto.class).getResultList();
-
     }
 }
